Easily specify colors and turn off filtering
Now it is easy to just specify which colors to use for mw onset, self report, and mw onset 2 at the top of the onset_vs_self_report script. There is also now a flag in the preprocess script so that it is easy to turn off the functionality that filters out MW Onset events too close to page end.